==> real_state_new/1_location_price.ipynb <==
# 将百城2010年6月到2016年3月的房价整合到包含坐标的excel表中

import numpy as np
from pandas import Series, DataFrame
import pandas as pd

result = pd.read_csv("../data/cities_coordinate.csv", encoding = 'gb18030')

for i in range(201006, 201013):
    filename = "../data/average/" + str(i) + '.csv'
    data = pd.read_csv(filename, encoding = 'gb18030')
    data = data.ix[1:,[0,1]]
    data.index = range(len(data))
    data.columns = ['city', i]
    result = pd.merge(result, data)

for i in range(2011, 2016):
    for j in range(1, 13):
        date = i * 100 + j
        filename = "../data/average/" + str(date) + '.csv'
        data = pd.read_csv(filename, encoding = 'gb18030')
        data = data.ix[1:,[0,1]]
        data.index = range(len(data))
        data.columns = ['city', date]
        result = pd.merge(result, data)

for i in range(201601, 201604):
    filename = "../data/average/" + str(i) + '.csv'
    data = pd.read_csv(filename, encoding = 'gb18030')
    data = data.ix[1:,[0,1]]
    data.index = range(len(data))
    data.columns = ['city', i]
    result = pd.merge(result, data)

result.to_csv('../data/100cities_location_price.csv',encoding = 'gb18030', index=False)
