==> real_state_new/2_distance.ipynb <==
# 计算长三角区域各城市之间的距离及倒数

import numpy as np
from pandas import Series, DataFrame
import pandas as pd

location = pd.read_csv("../data/changsanjiao/location_changsanjiao.csv", encoding = 'gb18030')

location

from math import *

# input Lat_A 纬度A
# input Lng_A 经度A
# input Lat_B 纬度B
# input Lng_B 经度B
# output distance 距离(km)
def calcDistance(Lat_A, Lng_A, Lat_B, Lng_B):
    if Lat_A == Lat_B:
        distance = 0
    else:
        ra = 6378.140  # 赤道半径 (km)
        rb = 6356.755  # 极半径 (km)
        flatten = (ra - rb) / ra  # 地球扁率
        rad_lat_A = radians(Lat_A)
        rad_lng_A = radians(Lng_A)
        rad_lat_B = radians(Lat_B)
        rad_lng_B = radians(Lng_B)
        pA = atan(rb / ra * tan(rad_lat_A))
        pB = atan(rb / ra * tan(rad_lat_B))
        xx = acos(sin(pA) * sin(pB) + cos(pA) * cos(pB) * cos(rad_lng_A - rad_lng_B))
        c1 = (sin(xx) - xx) * (sin(pA) + sin(pB)) ** 2 / cos(xx / 2) ** 2
        c2 = (sin(xx) + xx) * (sin(pA) - sin(pB)) ** 2 / sin(xx / 2) ** 2
        dr = flatten / 8 * (c1 - c2)
        distance = ra * (xx + dr)
    return distance

Lat_A=32.060255; Lng_A=118.796877 # 南京
Lat_B=39.904211; Lng_B=116.407395 # 北京
distance=calcDistance(Lat_A,Lng_A,Lat_B,Lng_B)
print('(Lat_A, Lng_A)=({0:10.3f},{1:10.3f})'.format(Lat_A,Lng_A))
print('(Lat_B, Lng_B)=({0:10.3f},{1:10.3f})'.format(Lat_B,Lng_B))
print('Distance={0:10.3f} km'.format(distance))

all_cities_list = list()

for i in range(0, 19):     
    city_list = list()
    
    for j in range(0, 19):
        temp_location = list(location.ix[i, 1:3]) + (list(location.ix[j, 1:3]))
        distance = calcDistance(*temp_location)
        city_list.append(distance)
        
    all_cities_list.append(city_list)

result = DataFrame(all_cities_list)

result.index = location["city"]
result.columns = location["city"]
result.to_csv("../data/changsanjiao/distance_changsanjiao.csv", encoding = 'gb18030')

result_inverse = 1.0 / result
result_inverse.to_csv("../data/changsanjiao/distance_changsanjiao_inverse.csv", encoding = 'gb18030')
