==> real_state_new/3_reverse.ipynb <==
# 取出长三角区域的城市样本，保留价格数据，并将价格按日期从近到远排列

import numpy as np
import pandas as pd 
from pandas import Series,DataFrame

location = pd.read_csv("../data/changsanjiao/location_changsanjiao.csv", encoding = 'gb18030')
cities = location["city"]

data = pd.read_csv('../data/100cities_location_price.csv',encoding ='gb18030')
data = data.set_index("city")
data = data.ix[cities]
data.columns = range(data.shape[1])
data = data.ix[:, 4:]
data.columns = range(data.shape[1])
data.index = range(data.shape[0])

raw_df = np.arange(1330).reshape(19, 70)
df = DataFrame(raw_df, index = range(0, 19), columns = range(0, 70))

for i in range(0, 70):
    df[i] = data[69-i]

df.shape

df

df.to_csv('../data/changsanjiao/price_reverse_changsanjiao.csv', encoding='gb18030', index=None)
