==> real_state_new/4_product.ipynb <==
# 计算距离倒数矩阵与价格矩阵的乘积

import numpy as np
from pandas import Series, DataFrame
import pandas as pd

dis = pd.read_csv("../data/changsanjiao/distance_changsanjiao_inverse.csv", encoding = 'gb18030')
dis = dis.ix[:, 1:]
dis[dis == np.inf] = 0

dis.shape

dis.head()

price = pd.read_csv("../data/changsanjiao/price_changsanjiao_reverse.csv", encoding = 'gb18030')

product = np.dot(dis, price)
product = DataFrame(product)

product.to_csv("../data/changsanjiao/product_changsanjiao.csv", encoding = 'gb18030', index=None)
