==> real_state_new/5_final_data.ipynb <==
# 按照论文中的要求，将长三角19个城市的价格、乘积数据及滞后项整理出来

import numpy as np
from pandas import Series, DataFrame
import pandas as pd

price = pd.read_csv("../data/changsanjiao/price_changsanjiao_reverse.csv", encoding = 'gb18030')
product = pd.read_csv("../data/changsanjiao/product_changsanjiao.csv", encoding = 'gb18030')

i = 0
variables_0 = DataFrame()

price_raw_0 = price.xs(i)
price_i = price.xs(i)
price_i = price_i.diff()[1:]
price_i = -1 * price_i
price_i.index = range(len(price_i))

for j in range(0, 5):
    temp = price_i[j:(65+j)]
    temp.index = range(len(temp))
    variables_0[j] = temp

product_raw = product.xs(i)
product_i = product.xs(i)
product_i = product_i.diff()[1:]
product_i = -1 * product_i
product_i.index = range(len(product_i))

for j in range(0, 4):
    temp = product_i[j+1:(66+j)]
    temp.index = range(len(temp))
    variables_0[j+5] = temp

price_product = price_raw_0[1:] - product_raw[1:]
price_product.index = range(len(price_product))
variables_0[9] = price_product[:65]

variables_0.columns = ['t', 't-1', 't-2', 't-3', 't-4', 'p-1', 'p-2', 'p-3', 'p-4', 't-p']

filename = '../data/changsanjiao/regression_0/' + str(i) + '.csv'
variables_0.to_csv(filename, encoding = 'gb18030', index=None)

variables_0 = variables_0[['t', 't-1', 't-2', 't-3', 't-4']]
variables_0.columns = ['t0', 't0-1', 't0-2', 't0-3', 't0-4']

for i in range(1, 19):
    variables = DataFrame()

    price_raw = price.xs(i)
    price_i = price.xs(i)
    price_i = price_i.diff()[1:]
    price_i = -1 * price_i
    price_i.index = range(len(price_i))

    for j in range(0, 5):
        temp = price_i[j:(65+j)]
        temp.index = range(len(temp))
        variables[j] = temp

    product_raw = product.xs(i)
    product_i = product.xs(i)
    product_i = product_i.diff()[1:]
    product_i = -1 * product_i
    product_i.index = range(len(product_i))

    for j in range(0, 4):
        temp = product_i[j+1:(66+j)]
        temp.index = range(len(temp))
        variables[j+5] = temp

    price_product = price_raw[1:] - product_raw[1:]
    price_product.index = range(len(price_product))
    variables[9] = price_product[:65]
    
    price_price_0 = price_raw[1:] - price_raw_0[1:]
    price_price_0.index = range(len(price_price_0))
    variables[10] = price_price_0[:65]    

    variables.columns = ['t', 't-1', 't-2', 't-3', 't-4', 'p-1', 'p-2', 'p-3', 'p-4', 't-p', 't-t0']
    variables = pd.concat([variables, variables_0], axis=1)
    
    filename = '../data/changsanjiao/regression_0/' + str(i) + '.csv'
    variables.to_csv(filename, encoding = 'gb18030', index=None)
