==> real_state_new/8_matrix.ipynb <==
import numpy as np
from pandas import Series, DataFrame
import pandas as pd
from numpy.linalg import inv
import math

# 矩阵C
coe = pd.read_csv("../data/changsanjiao/regression_0/regression_coe_0.csv", encoding= 'gb18030')
coe = coe.drop("names", axis = 1)
for i in range(11,16):
    c = np.zeros([19,19])
    c[0] = coe.ix[i]
    c = c.T
    c = DataFrame(c)
    filename = "../data/changsanjiao/regression_0/C/C" + str(i-11) + ".csv"
    c.to_csv(filename, encoding = 'gb18030', index=None)

# C0 -> R

c0 = pd.read_csv("../data/changsanjiao/regression_0/C/C0.csv", encoding = 'gb18030')
r = DataFrame(inv(np.eye(19) - c0))
r.to_csv("../data/changsanjiao/regression_0/R/R.csv", encoding = 'gb18030', index=None)

# 矩阵A
coe = pd.read_csv("../data/changsanjiao/regression_0/regression_coe_0.csv", encoding = 'gb18030')
coe = coe.drop("names", axis = 1)
for i in range(5, 9):
    diag = DataFrame(np.diag(coe.ix[i]))
    filename = "../data/changsanjiao/regression_0/A/A" + str(i-4) + ".csv"
    diag.to_csv(filename, encoding = 'gb18030', index=None)

# S -> G
s = pd.read_csv("../data/changsanjiao/distance_changsanjiao_inverse.csv", encoding = 'gb18030')
s = s.drop("city", axis = 1)
b = pd.read_csv("../data/changsanjiao/regression_0/regression_coe_0.csv", encoding = 'gb18030')
b = b.drop("names", axis = 1)

for i in range(1, 5):
    df = DataFrame()
    b_vector = b.ix[i]
    for j in range(len(b_vector)):
        bs = b_vector[j] * s.ix[j]
        df[j] = bs
    df = df.T
    df.columns = range(19)
    filename = "../data/changsanjiao/regression_0/G/G" + str(i) + ".csv"
    df.to_csv(filename, encoding = 'gb18030', index=None)

# 矩阵H
s = pd.read_csv("../data/changsanjiao/distance_changsanjiao_inverse.csv", encoding = 'gb18030')
s = s.drop("city",axis = 1)
f = pd.read_csv("../data/changsanjiao/regression_0/regression_coe_0.csv", encoding = 'gb18030')
f = f.drop("names",axis = 1)
f_vector = f.ix[9]

h1 = np.diag(f_vector)

h2 = np.diag(f.ix[10])

h3 = np.zeros([19,19])
h3[0] = f.ix[10]
h3 = h3.T

h4 = DataFrame()
for j in range(len(f_vector)):
    fs = f_vector[j]*s.ix[j]
    h4[j] = fs
h4 = h4.T
h4.columns = range(19)

h = h1 + h2 - h3 - h4

h.to_csv("../data/changsanjiao/regression_0/H/H.csv", encoding = 'gb18030',index=None)

# gama
r = pd.read_csv("../data/changsanjiao/regression_0/R/R.csv", encoding = 'gb18030')

for i in range(1, 2):
    a_filename = "../data/changsanjiao/regression_0/A/A" + str(i) + ".csv"
    a = pd.read_csv(a_filename, encoding = 'gb18030')
    g_filename = "../data/changsanjiao/regression_0/G/G" + str(i) + ".csv"
    g = pd.read_csv(g_filename, encoding = 'gb18030')
    c_filename = "../data/changsanjiao/regression_0/C/C" + str(i) + ".csv"
    c = pd.read_csv(c_filename, encoding = 'gb18030')
    ag = np.add(a, g)
    agc = np.add(ag, c)
    gama = DataFrame(np.dot(r, agc))
    gama_filename = "../data/changsanjiao/regression_0/gama/gama" + str(i) + ".csv"
    gama.to_csv(gama_filename, encoding = 'gb18030', index=None)

# pi
r = pd.read_csv("../data/changsanjiao/regression_0/R/R.csv", encoding = 'gb18030')
h = pd.read_csv("../data/changsanjiao/regression_0/H/H.csv", encoding = 'gb18030')
pi = DataFrame(np.dot(r, h))
pi.to_csv("../data/changsanjiao/regression_0/pi/pi.csv", encoding = 'gb18030', index=None)

# phi1
pi = pd.read_csv("../data/changsanjiao/regression_0/pi/pi.csv", encoding='gb18030')
gama1 = pd.read_csv("../data/changsanjiao/regression_0/gama/gama1.csv", encoding = 'gb18030')
phi1 = np.add(np.eye(19), pi)
phi1 = np.add(phi1, gama1)
phi1.to_csv("../data/changsanjiao/regression_0/phi/phi1.csv", encoding = 'gb18030', index=None)

# phi2
gama2 = pd.read_csv("../data/changsanjiao/regression_0/gama/gama2.csv", encoding = 'gb18030')
phi2 = np.subtract(gama2, gama1)
phi2.to_csv("../data/changsanjiao/regression_0/phi/phi2.csv", encoding = 'gb18030', index=None)

# phi3
gama3 = pd.read_csv("../data/changsanjiao/regression_0/gama/gama3.csv", encoding = 'gb18030')
phi3 = np.subtract(gama3, gama2)
phi3.to_csv("../data/changsanjiao/regression_0/phi/phi3.csv", encoding = 'gb18030', index=None)

#phi4
gama4 = pd.read_csv("../data/changsanjiao/regression_0/gama/gama4.csv", encoding = 'gb18030')
phi4 = np.subtract(gama4, gama3)
phi4.to_csv("../data/changsanjiao/regression_0/phi/phi4.csv", encoding = 'gb18030', index=None)

#phi5
phi5 = -1 * gama4
phi5.to_csv("../data/changsanjiao/regression_0/phi/phi5.csv", encoding = 'gb18030', index=None)

#phi = pd.read_csv("../data/changsanjiao/regression_0/phi/phi1.csv", encoding = 'gb18030')
# psi0
psi0 = np.eye(19)
psi0 = DataFrame(psi0)
psi0.to_csv("../data/changsanjiao/regression_0/psi/psi0.csv", encoding = 'gb18030', index=None)

# psi1
psi1 = np.dot(phi1, psi0)
psi1 = DataFrame(psi1)
psi1.to_csv("../data/changsanjiao/regression_0/psi/psi1.csv", encoding = 'gb18030', index=None)

# psi2
psi2 = np.add(np.dot(phi1, psi1), np.dot(phi2, psi0))
psi2 = DataFrame(psi2)
psi2.to_csv("../data/changsanjiao/regression_0/psi/psi2.csv", encoding = 'gb18030', index=None)

# psi3
psi3 = np.add(np.dot(phi1, psi2), np.dot(phi2, psi1))
psi3 = np.add(psi3, np.dot(phi3, psi0))
psi3 = DataFrame(psi3)
psi3.to_csv("../data/changsanjiao/regression_0/psi/psi3.csv", encoding = 'gb18030', index=None)

# psi4
psi4 = np.add(np.dot(phi1, psi3), np.dot(phi2, psi2))
psi4 = np.add(psi4, np.dot(phi3, psi1))
psi4 = np.add(psi4, np.dot(phi4, psi0))
psi4 = DataFrame(psi4)
psi4.to_csv("../data/changsanjiao/regression_0/psi/psi4.csv", encoding = 'gb18030', index=None)


#psi5 - psi19
for i in range(5, 20):
    a1 = "../data/changsanjiao/regression_0/psi/psi" + str(i-1) +".csv"
    b1 = "../data/changsanjiao/regression_0/psi/psi" + str(i-2) +".csv"
    c1 = "../data/changsanjiao/regression_0/psi/psi" + str(i-3) +".csv"
    d1 = "../data/changsanjiao/regression_0/psi/psi" + str(i-4) +".csv"
    e1 = "../data/changsanjiao/regression_0/psi/psi" + str(i-5) +".csv"
    a = pd.read_csv(a1, encoding = 'gb18030')
    b = pd.read_csv(b1, encoding = 'gb18030')
    c = pd.read_csv(c1, encoding = 'gb18030')
    d = pd.read_csv(d1, encoding = 'gb18030')
    e = pd.read_csv(e1, encoding = 'gb18030')
    psi = np.dot(phi1, a)
    psi = np.add(psi, np.dot(phi2, b))
    psi = np.add(psi, np.dot(phi3, c))
    psi = np.add(psi, np.dot(phi4, d))
    psi = np.add(psi, np.dot(phi5, e))
    filename = "../data/changsanjiao/regression_0/psi/psi" + str(i) + ".csv"
    psi = DataFrame(psi)
    psi.to_csv(filename, encoding = 'gb18030', index=None)

# 脉冲响应函数g0(h) i=0时，表示中心城市对周边城市的影响
E = DataFrame(np.eye(19))
e = E[0]
sigma = pd.read_csv("../data/changsanjiao/regression_0/regression_resid_cov_0.csv", encoding = 'gb18030').drop("Unnamed: 0", axis=1)
sigma.columns = range(19)
sig = sigma.ix[0, 0]
r = pd.read_csv("../data/changsanjiao/regression_0/R/R.csv", encoding = 'gb18030')

df = DataFrame()

for h in range(0, 20):
    filename = "../data/changsanjiao/regression_0/psi/psi" + str(h) + ".csv"
    psi = pd.read_csv(filename, encoding='gb18030')

    gih = np.dot(psi, r)
    gih = np.dot(gih, sigma)
    gih = np.dot(gih, e)
    gih = gih / math.sqrt(sig)

    df[h] = gih
    
filename = "../data/changsanjiao/regression_0/pulse/pulse.csv"
df.to_csv(filename, encoding = 'gb18030', index=None)  

df_dis = df.reindex([0,4,11,5,2,12,10,3,13,9,8,7,6,1,15,18,17,14,16])
df_dis.to_csv("../data/changsanjiao/regression_0/pulse/pulse_dis.csv", encoding = 'gb18030', index=None)  

df_dis_diff = df_dis.diff(axis=1).drop(0, axis=1)
df_dis_diff.columns = range(19)
df_dis_diff.index = range(19)
df_dis_diff.to_csv("../data/changsanjiao/regression_0/pulse/pulse_dis_diff.csv", encoding = 'gb18030', index=None)  

# 脉冲响应函数gi(h) i表示周边城市对中心城市的影响
E = DataFrame(np.eye(19))
r = pd.read_csv("../data/changsanjiao/regression_0/R/R.csv", encoding = 'gb18030')
sigma = pd.read_csv("../data/changsanjiao/regression_0/regression_resid_cov_0.csv", encoding = 'gb18030').drop("Unnamed: 0", axis=1)
sigma.columns = range(19)

df = np.zeros([19,20])
df = DataFrame(df)

for i in range(0, 19):
    df_i = DataFrame()
    e = E[i]
    sig = sigma.ix[i, i]
    
    for h in range(0, 20):
        filename = "../data/changsanjiao/regression_0/psi/psi" + str(h) + ".csv"
        psi = pd.read_csv(filename, encoding='gb18030')

        gih = np.dot(psi, r)
        gih = np.dot(gih, sigma)
        gih = np.dot(gih, e)
        gih = gih / math.sqrt(sig)

        df_i[h] = gih
    df.ix[i] = df_i.ix[0]
    
filename = "../data/changsanjiao/regression_0/pulse1/pulse.csv"
df.to_csv(filename, encoding = 'gb18030', index=None)  

df_dis = df.reindex([0,4,11,5,2,12,10,3,13,9,8,7,6,1,15,18,17,14,16])
df_dis.to_csv("../data/changsanjiao/regression_0/pulse1/pulse_dis.csv", encoding = 'gb18030', index=None)  

df_dis_diff = df_dis.diff(axis=1).drop(0, axis=1)
df_dis_diff.columns = range(19)
df_dis_diff.index = range(19)
df_dis_diff.to_csv("../data/changsanjiao/regression_0/pulse1/pulse_dis_diff.csv", encoding = 'gb18030', index=None)  

# plot
import matplotlib.pyplot as plt
from mpl_toolkits.mplot3d import Axes3D

for h in range(0, 20):
    filename = "../data/changsanjiao/regression_0/pulse/pulse" + str(h) + ".csv"
    df = pd.read_csv(filename, encoding = 'gb18030')
    x = range(19) * 19
    y = list()
    z = Series()
    for i in range(19):
        z = pd.concat([z, df[str(i)]], ignore_index=True)
        y = y + [i] * 19
        
    fig = plt.figure()
    ax = fig.add_subplot(111, projection='3d')
    X = x
    Y = y
    Z = z
    ax.plot_trisurf(X, Y, Z)
    figname = "../data/changsanjiao/regression_0/figure/pulse" + str(h) + ".jpg"
    fig.savefig(figname)
